Fixed compiler warnings
client/mysqltest.cc:
  Added cast
plugin/feedback/utils.cc:
  Added #ifdef
sql/sql_select.cc:
  Added cast
sql/sys_vars.h:
  Changed limits to int
support-files/compiler_warnings.supp:
  Added suppression

+++ client/mysqltest.cc
@@ -5144,7 +5144,7 @@ typedef struct
 
 static st_error global_error_names[] =
 {
-  { "<No error>", -1, "" },
+  { "<No error>", (uint) -1, "" },
 #include <mysqld_ername.h>
   { 0, 0, 0 }
 };
